requirements.txt
click==8.1.7
Flask==3.0.0
numpy==1.26.2
Pillow==10.0.0
requests==2.31.0
